// packages/core/src/Types/Events.ts
export enum Events {
  BOOT_CONFIG = 'boot:config:load',
  BOOT_PLUGINS = 'boot:plugins:load',
  EXEC_MAIL_PARSE = 'execute:mail:parse',
  AFTER_EXEC_MAIL_PARSE = 'after:execute:mail:parse',
  EXEC_RECEIPT_PARSE = 'execute:receipt:parse',
  AFTER_EXEC_RECEIPT_PARSE = 'after:execute:receipt:parse',
}